# server/docs/swagger.yaml
basePath: /api
definitions:
  auth.Request:
    properties:
      credential:
        type: string
    required:
    - credential
    type: object
  gin.H:
    additionalProperties: {}
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.PrivateMessage:
    properties:
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      isRead:
        type: boolean
      receiverID:
        type: integer
      receiverUser:
        type: integer
      senderID:
        type: integer
      senderUser:
        $ref: '#/definitions/models.User'
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
      username:
        type: string
    type: object
  response.ResJSON:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a chat platform.
  termsOfService: http://swagger.io/terms/
  title: Chat Platform API
  version: "1.0"
paths:
  /auth/google:
    post:
      consumes:
      - application/json
      description: Verify Google ID Token and login
      parameters:
      - description: Google ID Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResJSON'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResJSON'
      summary: Google Login
      tags:
      - auth
  /messages:
    get:
      consumes:
      - application/json
      description: Get all private messages
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PrivateMessage'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get all private messages
      tags:
      - messages
    post:
      consumes:
      - application/json
      description: Get all private messages
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PrivateMessage'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get all private messages
      tags:
      - messages
  /users:
    get:
      consumes:
      - application/json
      description: Get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get all users
      tags:
      - users
swagger: "2.0"